get rid of vendored api script
